refactor(filemanager): 拆出传输队列面板的卡片与元数据模块
- 新增 components/filemanager/：uploadQueueMeta（状态/阶段/分块元数据与折叠队列构建）、
  AutoFollowChunkGrid（自动跟随滚动的分块网格）、UploadQueueCard（单条传输任务卡片）
- FileUploadQueuePanel.tsx 558 -> 149 行
